+++ src/distributions.h
@@ -4,6 +4,7 @@
 #include <math.h>
 #include <stdint.h>
 
+#include "tensor.h"
 #include "types.h"
 
 #ifndef M_PI
@@ -15,7 +16,7 @@ u32 unif_rand(u32 max);
 
 u32 unif_rand_range(u32 min, u32 max);
 
-void shuffle_array(u16* array, size_t length);
+void shuffle_array(u16* array, u32 length);
 
 /************* GAUSSIAN ***********/
 // Returns a random number sampled from N(0,1)
